feat: refactor to create arena.h
